refactor: move api evaluation_context methods to api module
