+++ open_feature/open_feature_api.py
@@ -1,5 +1,6 @@
 import typing
 
+from open_feature.evaluation_context.evaluation_context import EvaluationContext
 from open_feature.exception.exceptions import GeneralError
 from open_feature.open_feature_client import OpenFeatureClient
 from open_feature.provider.metadata import Metadata
@@ -7,6 +8,8 @@ from open_feature.provider.no_op_provider import NoOpProvider
 from open_feature.provider.provider import AbstractProvider
 
 _provider: AbstractProvider = NoOpProvider()
+
+_evaluation_context = EvaluationContext()
 
 
 def get_client(
@@ -30,3 +33,15 @@ def get_provider() -> typing.Optional[AbstractProvider]:
 def get_provider_metadata() -> typing.Optional[Metadata]:
     global _provider
     return _provider.get_metadata()
+
+
+def api_evaluation_context() -> EvaluationContext:
+    global _evaluation_context
+    return _evaluation_context
+
+
+def set_api_evaluation_context(evaluation_context: EvaluationContext):
+    global _evaluation_context
+    if evaluation_context is None:
+        raise GeneralError(error_message="No api level evaluation context")
+    _evaluation_context = evaluation_context
